Moving converter to Json directory:
